requirements.txt
Pillow==7.1.2
PyMuPDF==1.17.0
